+++ app/src/main/AndroidManifest.xml
@@ -12,15 +12,15 @@
         android:theme="@style/Theme.Movie"
         tools:targetApi="31">
         <activity
-            android:name=".ui.CompletedActivity"
-            android:parentActivityName=".ui.BookingActivity"
+            android:name=".ui.completed.CompletedActivity"
+            android:parentActivityName=".ui.booking.BookingActivity"
             android:exported="false" />
         <activity
-            android:name=".ui.BookingActivity"
-            android:parentActivityName=".ui.MainActivity"
+            android:name=".ui.booking.BookingActivity"
+            android:parentActivityName=".ui.main.MainActivity"
             android:exported="false" />
         <activity
-            android:name=".ui.MainActivity"
+            android:name=".ui.main.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
