+++ app/src/main/java/woowacourse/movie/ui/main/MovieAdapter.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.ui
+package woowacourse.movie.ui.main
 
 import android.content.Context
 import android.view.LayoutInflater
